fix: rename grapher to parser
